refactor: feedback

+++ packages/trpc/server/routers/viewer/admin/watchlist/list.schema.ts
@@ -3,8 +3,8 @@ import { z } from "zod";
 import { WatchlistType, WatchlistSource } from "@calcom/prisma/enums";
 
 export const ZListWatchlistEntriesInputSchema = z.object({
-  limit: z.number().min(1).max(100).default(25),
-  offset: z.number().min(0).default(0),
+  limit: z.number().int().min(1).max(100).default(25),
+  offset: z.number().int().min(0).default(0),
   searchTerm: z.string().optional(),
   filters: z
     .object({
